+++ src/django_twc_toolbox/sentry.py
@@ -1,23 +1,28 @@
 from __future__ import annotations
 
+from typing import TYPE_CHECKING
+
 from django_twc_toolbox.conf import app_settings
 
+if TYPE_CHECKING:
+    from sentry_sdk._types import SamplingContext
 
-def sentry_traces_sampler(sampling_context):
+
+def sentry_traces_sampler(sampling_context: SamplingContext):
     if _should_disgard(sampling_context):
         return 0
 
     return app_settings.SENTRY_TRACES_RATE
 
 
-def sentry_profiles_sampler(sampling_context):
+def sentry_profiles_sampler(sampling_context: SamplingContext):
     if _should_disgard(sampling_context):
         return 0
 
     return app_settings.SENTRY_PROFILE_RATE
 
 
-def _should_disgard(sampling_context) -> bool:
+def _should_disgard(sampling_context: SamplingContext) -> bool:
     DISGARDED_METHODS = app_settings.SENTRY_DISGARDED_METHODS
     DISGARDED_PATHS = app_settings.SENTRY_DISGARDED_PATHS