+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project attempts to adhere to [Semantic Versioning](https://semver.org/
 
 ## [Unreleased]
 
+## [0.9.0]
+
 ### Added
 
 - Two new templatetag filters: `{{ variable|klass }}` and `{{ variable|class_name }}`.
@@ -131,7 +133,7 @@ Initial release!
 
 - Josh Thomas <josh@joshthomas.dev> (maintainer)
 
-[unreleased]: https://github.com/westerveltco/django-twc-toolbox/compare/v0.8.0...HEAD
+[unreleased]: https://github.com/westerveltco/django-twc-toolbox/compare/v0.9.0...HEAD
 [0.2.1]: https://github.com/westerveltco/django-twc-toolbox/releases/tag/v0.2.1
 [0.2.0]: https://github.com/westerveltco/django-twc-toolbox/releases/tag/v0.2.0
 [0.1.1]: https://github.com/westerveltco/django-twc-toolbox/releases/tag/v0.1.1
@@ -142,3 +144,4 @@ Initial release!
 [0.6.0]: https://github.com/westerveltco/django-twc-toolbox/releases/tag/v0.6.0
 [0.7.0]: https://github.com/westerveltco/django-twc-toolbox/releases/tag/v0.7.0
 [0.8.0]: https://github.com/westerveltco/django-twc-toolbox/releases/tag/v0.8.0
+[0.9.0]: https://github.com/westerveltco/django-twc-toolbox/releases/tag/v0.9.0